# openapi.yaml
openapi: 3.0.3
info:
  title: CatWAF API
  version: 1.0.1
  description: |
    CatWAF's control-plane API — everything the dashboard itself calls.
    There's no separate "public API" here; this is the real thing the
    frontend uses, documented so you can script against it or build your
    own tooling on top.

    **Scope of this document:** CatWAF has ~167 routes across 19 files.
    This spec covers the core groups in real depth — auth and two-factor
    login, WAF engine settings, custom rules, IP/geo lists, the security
    score, honeypots, sensitive-file protection, operational modes
    (panic/maintenance/config signing), and the settings, protection and
    operations layers — rather than a shallow stub for every endpoint.
    Routes not yet documented here still exist and work; see
    `backend/routes/` for the full list. Expanding this file's coverage is
    tracked in `ROADMAP.md`.

    **The settings layer is self-describing.** Rather than enumerating ~290
    fields here, `GET /api/settings/schema` returns the live schema the
    backend validates against and the dashboard builds its controls from —
    field names, types, defaults, permitted values and help text. That is
    the authoritative reference; `docs/settings.md` is generated from it.

    **One thing worth knowing:** the dashboard stats/traffic endpoints
    (`/api/stats`, `/api/traffic/*`) return real data — sourced from
    Coraza's own audit log via `services/requestLog.js`, not a demo feed —
...
servers:
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Obtained from POST /api/auth/login. Unauthenticated requests are
        still accepted for GET routes (treated as the built-in `guest`/
        viewer role) — the token only matters for anything that mutates
        state, which requires the `admin` role.
  responses:
    Unauthorized:
      description: Read-only role — log in as admin to modify this.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        detail: { type: string, example: 'Read-only role. Log in as admin to modify.' }
...
    AppliedRoute:
      type: object
      description: One route in CatWAF's generated region of the live Caddyfile.
      properties:
        name: { type: string }
        runtime: { type: string, nullable: true, description: 'Detected runtime, e.g. "PHP 8.3"' }
        containerName: { type: string, nullable: true }
        listenPort: { type: integer, description: Port CatWAF listens on for this application }
        upstream: { type: string, nullable: true, description: Address CatWAF forwards to once a request passes the WAF }
    DiscoveredApp:
      type: object
      description: >
        The safe projection of a discovered container. Raw environment and
        labels are deliberately absent — see backend/services/discovery/summary.js.
      properties:
        name: { type: string, description: Compose service name where there is one, otherwise the container name }
        containerName: { type: string }
        image: { type: string }
        running: { type: boolean }
        isWeb: { type: boolean, description: Whether this serves HTTP and is therefore routable }
        webServer: { type: string, nullable: true }
        webServerConfidence: { type: number, nullable: true }
        runtime: { type: string, nullable: true }
        framework: { type: string, nullable: true }
        port: { type: integer, nullable: true }
        published: { type: boolean, description: Published to the host — such a port bypasses CatWAF while it stays published }
        reachability: { type: string, nullable: true }
        networks: { type: array, items: { type: string } }
        phpConfidence: { type: number, nullable: true }
        fastcgiBackend: { type: string, nullable: true }
        fpmBackendFor: { type: array, nullable: true, items: { type: string } }
        incomplete: { type: boolean }
        evidence:
          type: object
          description: Why CatWAF classified it this way. Files, processes, ports and config directives — never secret values.
          properties:
            web: { type: array, items: { type: string } }
            php: { type: array, items: { type: string } }
            runtime: { type: array, items: { type: string } }
            server: { type: array, items: { type: string } }
    VerificationResult:
      type: object
      properties:
        name: { type: string }
        protected: { type: boolean, description: True only when a benign request was served AND a CRS payload was refused }
        reason: { type: string, nullable: true, description: Why the route is not protected }
        benignStatus: { type: integer, nullable: true }
        attackStatus: { type: integer, nullable: true }
        upstreamUnreachable: { type: boolean, description: 'CatWAF is listening but the application behind it is down — the site is broken, not merely unprotected' }
        wafBypassed: { type: boolean, description: 'A CRS payload reached the application, so Coraza is not inspecting this route' }
    Verification:
      type: object
      properties:
        results: { type: array, items: { $ref: '#/components/schemas/VerificationResult' } }
        protectedCount: { type: integer }
        allProtected: { type: boolean }
    ProtectResult:
      type: object
      properties:
        ok: { type: boolean }
        stage:
          type: string
          description: How far the pipeline got.
          enum: [discovery, plan, dry-run, apply, applied, unverified, protected]
        dry_run: { type: boolean }
        error: { type: string, nullable: true }
        message: { type: string, nullable: true }
        containers: { type: array, items: { $ref: '#/components/schemas/DiscoveredApp' } }
        web_apps: { type: array, items: { $ref: '#/components/schemas/DiscoveredApp' } }
        routes: { type: array, items: { $ref: '#/components/schemas/AppliedRoute' } }
        skipped:
          type: array
          description: Applications CatWAF found but could not route to, with diagnostics.
          items:
            type: object
            properties:
              name: { type: string }
              reason: { type: string }
              diagnostics: { type: array, items: { type: string } }
        backup: { type: string, nullable: true, description: Path the previous Caddyfile was copied to }
        reloaded: { type: boolean, nullable: true }
        reload_error: { type: string, nullable: true }
        verification: { allOf: [{ $ref: '#/components/schemas/Verification' }], nullable: true }
        protected:
          type: boolean
          description: >
            Derived solely from `verification.allProtected`. A generated,
            applied and cleanly reloaded configuration does not set this.
...
paths:
...
  /api/apps/protect:
    post:
      tags: [Applications]
      summary: Apply the generated configuration, reload, then verify with real traffic
      description: >
        Admin only, and audited. Rewrites CatWAF's generated region, backs up the
        previous Caddyfile, reloads, then proves the result with a real benign
        request and a real inert CRS payload against each route.


        Unlike `catwaf auto`, this never starts the proxy: that is a privileged
        host action. If Caddy is not running the response reports the routes as
        applied but not live, rather than trying to start services.
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                skip_verify:
                  type: boolean
                  default: false
                  description: Apply without verifying. `protected` is then false, because nothing was proven.
      responses:
        '200':
          description: The pipeline ran; read `stage`, `verification` and `protected` for the outcome.
          content: { application/json: { schema: { $ref: '#/components/schemas/ProtectResult' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { description: Read-only role. }
        '409': { description: Another pipeline run is already in progress. }
...
  /api/bans:
    get:
      tags: [Protection]
      summary: Every address currently being refused
      description: >
        The unified ban store. Behavioural banning, DNSBL hits, the challenge
        gate, community lists, the threat feed and the shared network all write
        here. The manual IP blocklist is separate and is not included.
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: source, in: query, schema: { type: string }, description: Filter to one ban source }
        - { name: limit, in: query, schema: { type: integer, default: 200, maximum: 1000 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
        - { name: include_expired, in: query, schema: { type: boolean, default: false } }
      responses:
        '200': { description: OK }
    post:
      tags: [Protection]
      summary: Ban an address or range by hand
      description: >
        Refuses a range covering the caller's own address — banning it would
        lock the administrator out of the dashboard. Omit seconds for a
        permanent ban.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target]
              properties:
                target: { type: string, description: IP address or CIDR range }
                seconds: { type: integer, nullable: true, description: 'Duration; null means permanent' }
                reason: { type: string }
      responses:
        '200': { description: Banned }
        '400': { description: 'Invalid target, or code SELF_BLOCK when the range covers the caller' }
  /api/bans/{id}:
    delete:
      tags: [Protection]
      summary: Lift one ban
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: Lifted }
        '404': { description: No such ban }
  /api/bans/lift:
    post:
      tags: [Protection]
      summary: Lift every ban on an address, across sources
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [target], properties: { target: { type: string }, source: { type: string, nullable: true } } } } }
      responses:
        '200': { description: OK }
  /api/bans/clear:
    post:
      tags: [Protection]
      summary: Lift every automatic ban
      description: The manual IP blocklist is unchanged.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/protect/test:
    post:
      tags: [Protection]
      summary: Run the real classification pipeline against a supplied address
      description: >
        Answers "why would this visitor be treated the way they are" without
        waiting for them to visit. Returns the verdict and every signal that
        contributed, including sources that were unavailable and why.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [ip], properties: { ip: { type: string }, uri: { type: string, default: '/' }, user_agent: { type: string } } } } }
      responses:
        '200': { description: OK }
        '400': { description: Not a valid IP address }
  /api/protect/behavior:
    get:
      tags: [Protection]
      summary: Dry run — who behavioural banning would catch right now
      description: >
        Read from the real request log with the configured thresholds. Worth
        checking before enabling the feature: a legitimate client that 404s a
        lot is indistinguishable from a scanner by this measure.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/challenge/preview:
    get:
      tags: [Protection]
      summary: The real challenge page, issued by the real issuer
      description: Returns HTML, not JSON. Fails with 400 when the gate is switched off.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK, content: { text/html: { schema: { type: string } } } }
        '400': { description: The challenge gate is off, so there is nothing to preview }
  /api/challenge/scope-test:
    post:
      tags: [Protection]
      summary: Would this request be challenged?
      description: Runs the real scoping rules — exemptions, paths, country, user agent — against a hypothetical request.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, properties: { ip: { type: string }, uri: { type: string }, user_agent: { type: string }, country: { type: string, nullable: true }, suspicious: { type: boolean } } } } }
      responses:
        '200': { description: OK }
  /api/intel/lists:
    get:
      tags: [Protection]
      summary: Subscribed community blocklists and their entry counts
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }
  /api/intel/probe/tradeoffs:
    get:
      tags: [Protection]
      summary: The costs of active client probing, stated verbatim
      description: >
        Client probing opens a TCP connection back to the visitor. Some
        networks treat that as port scanning and will file an abuse report.
        This text is surfaced next to the toggle so enabling it is a decision
        rather than a surprise.
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: OK }

  # ─── Operations ────────────────────────────────────────────────────────
...